make abbr tags blocks so element title is visible when hovering over li
Another minor commit to your awesome design. Currently the abbr's title
is only visible when hovering over the abbr tag (which only occupies
part of the parent li/a tags).  These changes will cause the a/abbr
tags to occupy 100% of the li space so the title tag can be visible
when hovering over any part of the li, not just the abbr tag.